+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2023-12-17
+
 ### Added
 
 - "Name reference casing should match declaration" diagnostic.
@@ -34,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Document outline
 - Diagnostics
 
-[Unreleased]: https://github.com/cirras/vscode-eoplus/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/cirras/vscode-eoplus/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/cirras/vscode-eoplus/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/cirras/vscode-eoplus/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/cirras/vscode-eoplus/releases/tag/v1.0.0